requirements.txt
matplotlib==3.0.2
numpy==1.16.0
Pillow==5.4.1
torch==1.0.1.post2
torchvision==0.2.2.post3
